# docs/api/openapi/cordum-rest.yaml
openapi: 3.0.3
info:
  title: Cordum REST API
  version: 1.0.0
  description: |
    Safety-first agent orchestration platform. Cordum provides a control plane
    for submitting, scheduling, and governing AI agent jobs with built-in safety
    policies, approval workflows, dead-letter queues, and audit logging.

    ## Authentication
    All protected endpoints require either an API key (`X-API-Key` header) or a
    Bearer JWT token obtained from the login endpoint. Public endpoints are
    explicitly marked and require no credentials.

    ## Multi-Tenancy
    Protected routes require the `X-Tenant-ID` header to scope requests to the
    correct tenant partition.
  contact:
    name: Cordum Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8081
    description: Development
...
security:
  - apiKey: []
  - bearerAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /api/v1/locks/release:
    post:
      operationId: releaseLock
      summary: Release a distributed lock
      tags: [Locks]
      parameters:
        - $ref: "#/components/parameters/TenantID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LockRequest"
      responses:
        "200":
          description: Lock released
          content:
            application/json:
              schema:
                type: object
                properties:
                  lock:
                    $ref: "#/components/schemas/Lock"
                  released:
                    type: boolean
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
...
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key issued by the Cordum admin.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/auth/login.

  # =========================================================================
  # Parameters
  # =========================================================================
  parameters:
    TenantID:
      name: X-Tenant-ID
      in: header
      required: true
      description: Tenant isolation header (required on all protected routes).
      schema:
        type: string
...
  responses:
...
    Unauthorized:
      description: Unauthorized — missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
        status:
          type: integer
          description: HTTP status code
      required: [error, status]
...
    Lock:
      type: object
      properties:
        resource:
          type: string
        mode:
          type: string
          enum: [EXCLUSIVE, SHARED]
        owners:
          type: array
          items:
            type: string
        updated_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
    LockRequest:
      type: object
      properties:
        resource:
          type: string
        owner:
          type: string
        mode:
          type: string
          enum: [EXCLUSIVE, SHARED]
        ttl_ms:
          type: integer
          description: Lock time-to-live in milliseconds
      required: [resource, owner]

    # --- Packs ------------------------------------------------------------
...